+++ project/src/components/movieTabs/movieTabs.ts
